requirements.txt
openai>=1.40.0
python-dotenv>=1.0.0
streamlit>=1.28.0
azure-ai-projects>=1.0.0
azure-identity>=1.15.0
